git: remote source provider picker command
fixes #102394
