+++ extensions/git/src/commands.ts
@@ -6,10 +6,10 @@
 import { lstat, Stats } from 'fs';
 import * as os from 'os';
 import * as path from 'path';
-import { commands, Disposable, LineChange, MessageOptions, OutputChannel, Position, ProgressLocation, QuickPickItem, Range, SourceControlResourceState, TextDocumentShowOptions, TextEditor, Uri, ViewColumn, window, workspace, WorkspaceEdit, WorkspaceFolder, TimelineItem, env, QuickPick } from 'vscode';
+import { commands, Disposable, LineChange, MessageOptions, OutputChannel, Position, ProgressLocation, QuickPickItem, Range, SourceControlResourceState, TextDocumentShowOptions, TextEditor, Uri, ViewColumn, window, workspace, WorkspaceEdit, WorkspaceFolder, TimelineItem, env } from 'vscode';
 import TelemetryReporter from 'vscode-extension-telemetry';
 import * as nls from 'vscode-nls';
-import { Branch, GitErrorCodes, Ref, RefType, Status, CommitOptions, RemoteSourceProvider, RemoteSource } from './api/git';
+import { Branch, GitErrorCodes, Ref, RefType, Status, CommitOptions, RemoteSourceProvider } from './api/git';
 import { ForcePushMode, Git, Stash } from './git';
 import { Model } from './model';
 import { Repository, Resource, ResourceGroupType } from './repository';
@@ -18,8 +18,8 @@ import { fromGitUri, toGitUri, isGitUri } from './uri';
 import { grep, isDescendant, pathEquals } from './util';
 import { Log, LogLevel } from './log';
 import { GitTimelineItem } from './timelineProvider';
-import { throttle, debounce } from './decorators';
 import { ApiRepository } from './api/api1';
+import { pickRemoteSource } from './remoteSource';
 
 const localize = nls.loadMessageBundle();
 
@@ -238,72 +238,6 @@ interface PushOptions {
 	pushType: PushType;
 	forcePush?: boolean;
 	silent?: boolean;
-}
-
-async function getQuickPickResult<T extends QuickPickItem>(quickpick: QuickPick<T>): Promise<T | undefined> {
-	const result = await new Promise<T | undefined>(c => {
-		quickpick.onDidAccept(() => c(quickpick.selectedItems[0]));
-		quickpick.onDidHide(() => c(undefined));
-		quickpick.show();
-	});
-
-	quickpick.hide();
-	return result;
-}
-
-class RemoteSourceProviderQuickPick {
-
-	private quickpick: QuickPick<QuickPickItem & { remoteSource?: RemoteSource }>;
-
-	constructor(private provider: RemoteSourceProvider) {
-		this.quickpick = window.createQuickPick();
-		this.quickpick.ignoreFocusOut = true;
-
-		if (provider.supportsQuery) {
-			this.quickpick.placeholder = localize('type to search', "Repository name (type to search)");
-			this.quickpick.onDidChangeValue(this.onDidChangeValue, this);
-		} else {
-			this.quickpick.placeholder = localize('type to filter', "Repository name");
-		}
-	}
-
-	@debounce(300)
-	onDidChangeValue(): void {
-		this.query();
-	}
-
-	@throttle
-	async query(): Promise<void> {
-		this.quickpick.busy = true;
-
-		try {
-			const remoteSources = await this.provider.getRemoteSources(this.quickpick.value) || [];
-
-			if (remoteSources.length === 0) {
-				this.quickpick.items = [{
-					label: localize('none found', "No remote repositories found."),
-					alwaysShow: true
-				}];
-			} else {
-				this.quickpick.items = remoteSources.map(remoteSource => ({
-					label: remoteSource.name,
-					description: remoteSource.description || (typeof remoteSource.url === 'string' ? remoteSource.url : remoteSource.url[0]),
-					remoteSource
-				}));
-			}
-		} catch (err) {
-			this.quickpick.items = [{ label: localize('error', "$(error) Error: {0}", err.message), alwaysShow: true }];
-			console.error(err);
-		} finally {
-			this.quickpick.busy = false;
-		}
-	}
-
-	async pick(): Promise<RemoteSource | undefined> {
-		this.query();
-		const result = await getQuickPickResult(this.quickpick);
-		return result?.remoteSource;
-	}
 }
 
 export class CommandCenter {
@@ -527,51 +461,10 @@ export class CommandCenter {
 	@command('git.clone')
 	async clone(url?: string, parentPath?: string): Promise<void> {
 		if (!url) {
-			const quickpick = window.createQuickPick<(QuickPickItem & { provider?: RemoteSourceProvider, url?: string })>();
-			quickpick.ignoreFocusOut = true;
-
-			const providers = this.model.getRemoteProviders()
-				.map(provider => ({ label: (provider.icon ? `$(${provider.icon}) ` : '') + localize('clonefrom', "Clone from {0}", provider.name), alwaysShow: true, provider }));
-
-			quickpick.placeholder = providers.length === 0
-				? localize('provide url', "Provide repository URL")
-				: localize('provide url or pick', "Provide repository URL or pick a repository source.");
-
-			const updatePicks = (value?: string) => {
-				if (value) {
-					quickpick.items = [{
-						label: localize('repourl', "Clone from URL"),
-						description: value,
-						alwaysShow: true,
-						url: value
-					},
-					...providers];
-				} else {
-					quickpick.items = providers;
-				}
-			};
-
-			quickpick.onDidChangeValue(updatePicks);
-			updatePicks();
-
-			const result = await getQuickPickResult(quickpick);
-
-			if (result) {
-				if (result.url) {
-					url = result.url;
-				} else if (result.provider) {
-					const quickpick = new RemoteSourceProviderQuickPick(result.provider);
-					const remote = await quickpick.pick();
-
-					if (remote) {
-						if (typeof remote.url === 'string') {
-							url = remote.url;
-						} else if (remote.url.length > 0) {
-							url = await window.showQuickPick(remote.url, { ignoreFocusOut: true, placeHolder: localize('pick url', "Choose a URL to clone from.") });
-						}
-					}
-				}
-			}
+			url = await pickRemoteSource(this.model, {
+				providerLabel: provider => localize('clonefrom', "Clone from {0}", provider.name),
+				urlLabel: localize('repourl', "Clone from URL")
+			});
 		}
 
 		if (!url) {
@@ -2146,52 +2039,10 @@ export class CommandCenter {
 
 	@command('git.addRemote', { repository: true })
 	async addRemote(repository: Repository): Promise<string | undefined> {
-		const quickpick = window.createQuickPick<(QuickPickItem & { provider?: RemoteSourceProvider, url?: string })>();
-		quickpick.ignoreFocusOut = true;
-
-		const providers = this.model.getRemoteProviders()
-			.map(provider => ({ label: (provider.icon ? `$(${provider.icon}) ` : '') + localize('addfrom', "Add remote from {0}", provider.name), alwaysShow: true, provider }));
-
-		quickpick.placeholder = providers.length === 0
-			? localize('provide url', "Provide repository URL")
-			: localize('provide url or pick', "Provide repository URL or pick a repository source.");
-
-		const updatePicks = (value?: string) => {
-			if (value) {
-				quickpick.items = [{
-					label: localize('addFrom', "Add remote from URL"),
-					description: value,
-					alwaysShow: true,
-					url: value
-				},
-				...providers];
-			} else {
-				quickpick.items = providers;
-			}
-		};
-
-		quickpick.onDidChangeValue(updatePicks);
-		updatePicks();
-
-		const result = await getQuickPickResult(quickpick);
-		let url: string | undefined;
-
-		if (result) {
-			if (result.url) {
-				url = result.url;
-			} else if (result.provider) {
-				const quickpick = new RemoteSourceProviderQuickPick(result.provider);
-				const remote = await quickpick.pick();
-
-				if (remote) {
-					if (typeof remote.url === 'string') {
-						url = remote.url;
-					} else if (remote.url.length > 0) {
-						url = await window.showQuickPick(remote.url, { ignoreFocusOut: true, placeHolder: localize('pick url', "Choose a URL to clone from.") });
-					}
-				}
-			}
-		}
+		const url = await pickRemoteSource(this.model, {
+			providerLabel: provider => localize('addfrom', "Add remote from {0}", provider.name),
+			urlLabel: localize('addFrom', "Add remote from URL")
+		});
 
 		if (!url) {
 			return;
